so many machines, so little development

+++ src/test/java/autograder/filehandling/TestSubmissionReader.java
@@ -1,12 +1,9 @@
 package autograder.filehandling;
 
-import static org.junit.Assert.fail;
-
 import java.io.File;
 import java.io.IOException;
 
 import org.apache.commons.io.FileUtils;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -16,22 +13,23 @@ public class TestSubmissionReader {
 
 	@Before
 	public void setup() {
-		new File(Constants.SUBMISSIONS).mkdir();
-	}
-	
-	@After
-	public void tearDown() {
 		File subDir = new File(Constants.SUBMISSIONS);
-		try {
-			FileUtils.cleanDirectory(subDir);
-		} catch (IOException e) {
-			e.printStackTrace();
+		if(subDir.exists()) {
+			try {
+				FileUtils.cleanDirectory(subDir);
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		} else {
+			subDir.mkdirs();
 		}
 	}
 	
 	@Test
 	public void test() {
-		fail("Not yet implemented");
+		SubmissionReader sr = new SubmissionReader();
+		String filePath = getClass().getClassLoader().getResource("submissions.zip").getPath();
+		sr.unzipSubmissions(filePath);
 	}
 
 }
